backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pymongo==4.10.1
pydantic==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
groq==0.13.0
python-multipart==0.0.12
